backlight: qcom-wled: Use sink_addr for sync toggle
WLED3_SINK_REG_SYNC is, according to the name, a sink register offset.
Therefore, use the sink address as base.

This fixes the sync toggle on wled4, which can be
observed by the fact that adjusting brightness works.

It has no effect on wled3 because sink and ctrl base addresses are the same.
This allows adjusting the brightness without
having to disable then enable the module.

Signed-off-by: Obeida Shamoun <oshmoun100@googlemail.com>
Signed-off-by: Konrad Dybcio <konradybcio@gmail.com>