rpmsg: smd: Add quirk to initiate channel open from AP
RPM Firmware on variety of newer SoCs such as MSM8917 (also likely
MSM8937, MSM8940, MSM8952), MSM8953 and on some MSM8916 devices) doesn't
initiate opening of the SMD channel. This doesn't allow probing of
smd-rpm driver on such devices because there is a check that requires
RPM side of the channel to be opened first.

This commit adds a device-tree property that signals SMD driver that we
should open channel even when remote state is not OPENING or OPENED.

Signed-off-by: Vladimir Lypak <vladimir.lypak@gmail.com>