dt-bindings: display/msm: add MDSS/DPU support for SM6115
Add DPU and MDSS schemas to describe MDSS and DPU blocks on the Qualcomm
SM6115 platform.
Configuration for DSI/PHY is shared with QCM2290 so compatibles are reused.
Lack of dsi phy supply in example is intended 
due to fact on qcm2290, sm6115 and sm6125 
this phy is supplied via power domain, not regulator.